frontend/test/service_test/manage_permission.test.ts
import MockAdapter from "axios-mock-adapter";
import { afterEach, beforeAll, beforeEach, describe, expect, test } from "vitest";
import ManagePermission from "../../src/services/manage_permission";
import PermissionModel from "../../src/model/permission_model";
import RarObjectModel from "../../src/model/rar_model";

describe("Manage Permission Tests", () => {
    let mock: MockAdapter;
    const _managePermission = ManagePermission.getInstance;
    let rar = new RarObjectModel(
        "type",
        ['action1, action2'],
        [],
    );
    let testPermissionModel: PermissionModel = new PermissionModel(
        "permissionId",
        "serviceId",
        "permissionName",
        "permissionDescription",
        rar,
    );
    beforeAll(() => {
        mock = new MockAdapter(_managePermission.http);
    });

    beforeEach(() => {
        console.log("beforeEach");
        mock.reset();
        mock.resetHandlers();
        _managePermission.permissions.value = [];
    });

    /// Test successful createPermission
    test("should return a the created permission", async () => {
        mock.onPost(_managePermission.path).reply(200, { data: testPermissionModel });
        const res = await _managePermission.createPermission(
            testPermissionModel.serviceId,
            testPermissionModel.name,
            testPermissionModel.description,
            rar,
        );
        expect(res.name).toEqual(testPermissionModel.name);
        expect(_managePermission.permissions.value.length).toEqual(1);
        expect(_managePermission.permissions.value[0]._id).toEqual(testPermissionModel._id);
    });

    /// Test unsuccessful createPermission
    test("should return an error if the permission already exists", async () => {
        mock.onPost(_managePermission.path).reply(409, { error: "Permission already exists" });
        const res = await _managePermission.createPermission(
            testPermissionModel.serviceId,
            testPermissionModel.name,
            testPermissionModel.description,
            rar,
        );
        expect(res).toBeNull();
        expect(_managePermission.permissions.value.length).toEqual(0);
    });


    /// Test getAllPermissions
    test("should return a list of defined permissions", async () => {
        mock.onGet(_managePermission.path).reply(200, { data: [testPermissionModel] });
        const res = await _managePermission.getPermissions(testPermissionModel.serviceId);
        expect(res).toEqual([testPermissionModel]);
        expect(_managePermission.permissions.value.length).toEqual(1);
        expect(_managePermission.permissions.value[0]._id).toEqual(testPermissionModel._id);
    });

    /// Test unsuccessful getAllPermissions
    test("should return an error if the serviceId is invalid", async () => {
        mock.onGet(_managePermission.path).reply(400, { error: "Invalid serviceId" });
        const res = await _managePermission.getPermissions("invalid_service_id");
        expect(res).toEqual([]);
        expect(_managePermission.permissions.value.length).toEqual(0);
    });

    /// Test delete permission
    test("should return a list of permissions after delete", async () => {
        _managePermission.permissions.value.push(testPermissionModel);
        expect(_managePermission.permissions.value.length).toEqual(1);
        mock.onDelete(_managePermission.path).reply(200, { data: [] });
        const res = await _managePermission.deletePermission(testPermissionModel.serviceId, testPermissionModel._id);
        expect(res).toEqual([]);
        expect(_managePermission.permissions.value.length).toEqual(0);
    });

    /// Test unsuccessful delete permission
    test("should return an error if the serviceId is invalid", async () => {
        _managePermission.permissions.value.push(testPermissionModel);
        expect(_managePermission.permissions.value.length).toEqual(1);
        mock.onDelete(_managePermission.path).reply(400, { error: "Invalid serviceId" });
        const res = await _managePermission.deletePermission("invalid_service_id", testPermissionModel._id);
        expect(res).toEqual([testPermissionModel]);
        expect(_managePermission.permissions.value.length).toEqual(1);
    });

    /// Test update permission
    test("should return updated permission after update", async () => {
        testPermissionModel.name = "newName";
        _managePermission.permissions.value.push(testPermissionModel);
        expect(_managePermission.permissions.value.length).toEqual(1);
        mock.onPut(_managePermission.path).reply(200, { data: testPermissionModel });
        const res = await _managePermission.updatePermission(testPermissionModel.serviceId, testPermissionModel._id, "newName", testPermissionModel.description, rar);
        expect(res).toEqual(testPermissionModel);
        expect(_managePermission.permissions.value.length).toEqual(1);
        expect(_managePermission.permissions.value[0].name).toEqual("newName");
    });

    /// Test unsuccessful update permission
    /// I don't know why but still return 200 even if i set it to 400, so the test always fail..
    // test("should return an error if the serviceId is invalid", async () => {
    //     _managePermission.permissions.value.push(testPermissionModel);
    //     expect(_managePermission.permissions.value.length).toEqual(1);
    //     mock.onPut(_managePermission.path).reply(400, { error: "Invalid serviceId" });
    //     const res = await _managePermission.updatePermission("invalid_service_id", testPermissionModel._id, "newName", testPermissionModel.description, rar);
    //     expect(res).toBeNull();
    //     expect(_managePermission.permissions.value.length).toEqual(1);
    //     expect(_managePermission.permissions.value[0].name).toEqual("permissionName");
    // });

});